# attendance-system/backup_before_restore_20260716_200909/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Attendance System API
  description: >
    API quản lý chấm công tập trung, tích hợp nhiều dòng máy (ZKTeco, Sunbeam/Timmy, Hikvision).
    Hệ thống chỉ lưu trữ raw log chấm công và đồng bộ nhân viên hai chiều — không tính công.
  version: "1.0.0"
servers:
  - url: /api/v1

# JWT Bearer Auth áp dụng cho toàn bộ API (trừ /auth/login được override bên dưới)
security:
  - BearerAuth: []
...
paths:
...
  /employees:
    post:
      tags: [employees]
      summary: Tạo nhân viên mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Nhân viên đã được tạo
    get:
      tags: [employees]
      summary: Danh sách nhân viên
      responses:
        '200':
          description: Danh sách nhân viên
  /employees/{id}:
    put:
      tags: [employees]
      summary: Sửa nhân viên
      parameters:
        - $ref: '#/components/parameters/EmployeeID'
      responses:
        '200':
          description: Nhân viên đã được cập nhật
    delete:
      tags: [employees]
      summary: Xoá nhân viên
      parameters:
        - $ref: '#/components/parameters/EmployeeID'
      responses:
        '204':
          description: Đã xoá
  /employees/import:
    post:
      tags: [employees]
      summary: Import danh sách nhân viên từ file Excel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import thành công
  /devices/{id}/sync-employees:
    post:
      tags: [sync]
      summary: Đẩy danh sách nhân viên active xuống máy chấm công
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '200':
          description: Kết quả đồng bộ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncHistory'
  /devices/{id}/sync-attendance:
    post:
      tags: [sync]
      summary: Đồng bộ log chấm công thủ công từ thiết bị
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Kết quả đồng bộ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncHistory'
  /attendance-logs:
    get:
      tags: [sync]
      summary: Truy vấn raw log chấm công
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: employee_code
          schema: { type: string }
        - in: query
          name: device_id
          schema: { type: string }
      responses:
        '200':
          description: Danh sách log chấm công
  /sync-history:
    get:
      tags: [sync]
      summary: Lịch sử đồng bộ
      parameters:
        - in: query
          name: device_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [success, failed, partial] }
      responses:
        '200':
          description: Danh sách lịch sử đồng bộ
  /sync-history/{id}/retry:
    post:
      tags: [sync]
      summary: Đồng bộ lại nếu lần trước thất bại
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Kết quả đồng bộ lại
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Lấy token từ POST /auth/login.
        Gửi kèm header: Authorization: Bearer <token>
  parameters:
    DeviceID:
      name: id
      in: path
      required: true
      schema: { type: string }
    EmployeeID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    CreateEmployeeRequest:
      type: object
      required: [employee_code, full_name]
      properties:
        employee_code: { type: string, example: "EMP001" }
        full_name: { type: string, example: "Nguyễn Văn A" }
        department_id: { type: string }
        card_no: { type: string }
        gender: { type: string, enum: [male, female, other] }
        date_of_birth: { type: string, format: date }
        phone: { type: string }
        email: { type: string }
        job_title: { type: string }
    SyncHistory:
      type: object
      properties:
        id: { type: integer }
        device_id: { type: string }
        sync_type: { type: string, enum: [employee, attendance, time_sync] }
        trigger_type: { type: string, enum: [manual, scheduled] }
        status: { type: string, enum: [success, failed, partial] }
        record_count: { type: integer }
        error_message: { type: string }
...